feat: :sparkles: cORS issue resolved
The CORS implemenation was previously done at the wrong entry point, adjusting the implementation to
the needed entry point fixed it
